applied linter suggestions

+++ contracts/lib/UniswapV2OracleLibrary.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.6.0;
 
-import '../interfaces/IUniswapV2Pair.sol';
+import "../interfaces/IUniswapV2Pair.sol";
 import "./UniswapV2Library.sol";
-import './FixedPoint.sol';
+import "./FixedPoint.sol";
 
 // library with helper methods for oracles that are concerned with computing average prices
 library UniswapV2OracleLibrary {
@@ -10,72 +10,111 @@ library UniswapV2OracleLibrary {
 
   // helper function that returns the current block timestamp within the range of uint32, i.e. [0, 2**32 - 1]
   function currentBlockTimestamp() internal view returns (uint32) {
-    return uint32(block.timestamp % 2 ** 32);
+    return uint32(block.timestamp % 2**32);
   }
 
   // produces the cumulative prices using counterfactuals to save gas and avoid a call to sync.
-  function currentCumulativePrices(
-    address pair
-  ) internal view returns (uint price0Cumulative, uint price1Cumulative, uint32 blockTimestamp) {
+  function currentCumulativePrices(address pair)
+    internal
+    view
+    returns (
+      uint256 price0Cumulative,
+      uint256 price1Cumulative,
+      uint32 blockTimestamp
+    )
+  {
     blockTimestamp = currentBlockTimestamp();
     price0Cumulative = IUniswapV2Pair(pair).price0CumulativeLast();
     price1Cumulative = IUniswapV2Pair(pair).price1CumulativeLast();
 
     // if time has elapsed since the last update on the pair, mock the accumulated price values
-    (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast) = IUniswapV2Pair(pair).getReserves();
+    (
+      uint112 reserve0,
+      uint112 reserve1,
+      uint32 blockTimestampLast
+    ) = IUniswapV2Pair(pair).getReserves();
     if (blockTimestampLast != blockTimestamp) {
       // subtraction overflow is desired
       uint32 timeElapsed = blockTimestamp - blockTimestampLast;
       // addition overflow is desired
       // counterfactual
-      price0Cumulative += uint(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
+      price0Cumulative += (
+        uint256(FixedPoint.fraction(reserve1, reserve0)._x) *
+        timeElapsed
+      );
       // counterfactual
-      price1Cumulative += uint(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
+      price1Cumulative += (
+        uint256(FixedPoint.fraction(reserve0, reserve1)._x) *
+        timeElapsed
+      );
     }
   }
 
   // produces the cumulative price using counterfactuals to save gas and avoid a call to sync.
   // only gets the first price
-  function currentCumulativePrice0(
-    address pair
-  ) internal view returns (uint price0Cumulative, uint32 blockTimestamp) {
+  function currentCumulativePrice0(address pair)
+    internal
+    view
+    returns (uint256 price0Cumulative, uint32 blockTimestamp)
+  {
     blockTimestamp = currentBlockTimestamp();
     price0Cumulative = IUniswapV2Pair(pair).price0CumulativeLast();
 
     // if time has elapsed since the last update on the pair, mock the accumulated price values
-    (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast) = IUniswapV2Pair(pair).getReserves();
+    (
+      uint112 reserve0,
+      uint112 reserve1,
+      uint32 blockTimestampLast
+    ) = IUniswapV2Pair(pair).getReserves();
     if (blockTimestampLast != blockTimestamp) {
       // subtraction overflow is desired
       uint32 timeElapsed = blockTimestamp - blockTimestampLast;
       // addition overflow is desired
       // counterfactual
-      price0Cumulative += uint(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
+      price0Cumulative += (
+        uint256(FixedPoint.fraction(reserve1, reserve0)._x) *
+        timeElapsed
+      );
     }
   }
 
   // produces the cumulative price using counterfactuals to save gas and avoid a call to sync.
   // only gets the second price
-  function currentCumulativePrice1(
-    address pair
-  ) internal view returns (uint price1Cumulative, uint32 blockTimestamp) {
+  function currentCumulativePrice1(address pair)
+    internal
+    view
+    returns (uint256 price1Cumulative, uint32 blockTimestamp)
+  {
     blockTimestamp = currentBlockTimestamp();
     price1Cumulative = IUniswapV2Pair(pair).price1CumulativeLast();
 
     // if time has elapsed since the last update on the pair, mock the accumulated price values
-    (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast) = IUniswapV2Pair(pair).getReserves();
+    (
+      uint112 reserve0,
+      uint112 reserve1,
+      uint32 blockTimestampLast
+    ) = IUniswapV2Pair(pair).getReserves();
     if (blockTimestampLast != blockTimestamp) {
       // subtraction overflow is desired
       uint32 timeElapsed = blockTimestamp - blockTimestampLast;
       // addition overflow is desired
       // counterfactual
-      price1Cumulative += uint(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
+      price1Cumulative += (
+        uint256(FixedPoint.fraction(reserve0, reserve1)._x) *
+        timeElapsed
+      );
     }
   }
 
   function getCurrentCumulativePrice(
-    address factory, address tokenIn, address tokenOut
-  ) internal view returns (uint priceCumulative, uint32 blockTimestamp) {
-    (address token0, address token1) = UniswapV2Library.sortTokens(tokenIn, tokenOut);
+    address factory,
+    address tokenIn,
+    address tokenOut
+  ) internal view returns (uint256 priceCumulative, uint32 blockTimestamp) {
+    (address token0, address token1) = UniswapV2Library.sortTokens(
+      tokenIn,
+      tokenOut
+    );
     address pair = UniswapV2Library.calculatePair(factory, token0, token1);
     if (tokenIn == token0) return currentCumulativePrice0(pair);
     return currentCumulativePrice1(pair);
@@ -93,8 +132,10 @@ library UniswapV2OracleLibrary {
   }
 
   function computeAverageAmountOut(
-    uint224 priceCumulativeStart, uint224 priceCumulativeEnd,
-    uint32 timeElapsed, uint256 amountIn
+    uint224 priceCumulativeStart,
+    uint224 priceCumulativeEnd,
+    uint32 timeElapsed,
+    uint256 amountIn
   ) internal pure returns (uint144 amountOut) {
     // overflow is desired.
     FixedPoint.uq112x112 memory priceAverage = computeAveragePrice(
